case-service/pyproject.toml
[project]
name = "case-service"
version = "0.1.0"
description = "Case and incident management service for SOC Forgex"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pymongo>=4.6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.2.1",
    "pyjwt>=2.8.0",
    "bcrypt>=4.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "mongomock>=4.1.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
